:bug: cast e.meta to decimal in src/server-extension/query/chart.ts

+++ src/server-extension/query/chart.ts
@@ -1,7 +1,7 @@
 export const maxBuy = `SELECT 
     date(e.timestamp), 
-    COALESCE(MAX(e.meta::bigint), 0) as value, 
-    COALESCE(AVG(e.meta::bigint), 0) as average, 
+    COALESCE(MAX(e.meta::decimal), 0) as value, 
+    COALESCE(AVG(e.meta::decimal), 0) as average, 
     COUNT(e.*) as count
     FROM event e 
     LEFT JOIN nft_entity ne on ne.id = e.nft_id 
@@ -11,7 +11,7 @@ export const maxBuy = `SELECT
 
   export const floorList = `SELECT 
     date(e.timestamp), 
-    COALESCE(MIN(e.meta::bigint), 0) as value,
+    COALESCE(MIN(e.meta::decimal), 0) as value,
     COUNT(e.*) as count
     FROM event e 
     LEFT JOIN nft_entity ne on ne.id = e.nft_id
